+++ src/server/bindings/objects.cpp
@@ -20,46 +20,46 @@
 
 extern NativeFunctionManager* g_native_function_manager;
 
-// Generated on 2014-09-27.
+// Generated on 2014-09-28.
 // Do not modify by hand, instead, look at /scripts/write_bindings.py.
 namespace samp {
 
-int CreateObject(int modelid, float X, float Y, float Z, float rX, float rY, float rZ, float DrawDistance) {
+int CreateObject(int modelid, double X, double Y, double Z, double rX, double rY, double rZ, double DrawDistance) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("CreateObject", "ifffffff", modelid, X, Y, Z, rX, rY, rZ, DrawDistance);
 }
 
-int AttachObjectToVehicle(int objectid, int vehicleid, float OffsetX, float OffsetY, float OffsetZ, float RotX, float RotY, float RotZ) {
+int AttachObjectToVehicle(int objectid, int vehicleid, double OffsetX, double OffsetY, double OffsetZ, double RotX, double RotY, double RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("AttachObjectToVehicle", "iiffffff", objectid, vehicleid, OffsetX, OffsetY, OffsetZ, RotX, RotY, RotZ);
 }
 
-int AttachObjectToObject(int objectid, int attachtoid, float OffsetX, float OffsetY, float OffsetZ, float RotX, float RotY, float RotZ, int SyncRotation) {
+int AttachObjectToObject(int objectid, int attachtoid, double OffsetX, double OffsetY, double OffsetZ, double RotX, double RotY, double RotZ, int SyncRotation) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("AttachObjectToObject", "iiffffffi", objectid, attachtoid, OffsetX, OffsetY, OffsetZ, RotX, RotY, RotZ, SyncRotation);
 }
 
-int AttachObjectToPlayer(int objectid, int playerid, float OffsetX, float OffsetY, float OffsetZ, float RotX, float RotY, float RotZ) {
+int AttachObjectToPlayer(int objectid, int playerid, double OffsetX, double OffsetY, double OffsetZ, double RotX, double RotY, double RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("AttachObjectToPlayer", "iiffffff", objectid, playerid, OffsetX, OffsetY, OffsetZ, RotX, RotY, RotZ);
 }
 
-int SetObjectPos(int objectid, float X, float Y, float Z) {
+int SetObjectPos(int objectid, double X, double Y, double Z) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("SetObjectPos", "ifff", objectid, X, Y, Z);
 }
 
-int GetObjectPos(int objectid, float* X, float* Y, float* Z) {
+int GetObjectPos(int objectid, double* X, double* Y, double* Z) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("GetObjectPos", "iFFF", objectid, X, Y, Z);
 }
 
-int SetObjectRot(int objectid, float RotX, float RotY, float RotZ) {
+int SetObjectRot(int objectid, double RotX, double RotY, double RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("SetObjectRot", "ifff", objectid, RotX, RotY, RotZ);
 }
 
-int GetObjectRot(int objectid, float* RotX, float* RotY, float* RotZ) {
+int GetObjectRot(int objectid, double* RotX, double* RotY, double* RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("GetObjectRot", "iFFF", objectid, RotX, RotY, RotZ);
 }
@@ -74,7 +74,7 @@ int DestroyObject(int objectid) {
   return g_native_function_manager->Invoke("DestroyObject", "i", objectid);
 }
 
-int MoveObject(int objectid, float X, float Y, float Z, float Speed, float RotX, float RotY, float RotZ) {
+int MoveObject(int objectid, double X, double Y, double Z, double Speed, double RotX, double RotY, double RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("MoveObject", "ifffffff", objectid, X, Y, Z, Speed, RotX, RotY, RotZ);
 }
@@ -89,27 +89,27 @@ int IsObjectMoving(int objectid) {
   return g_native_function_manager->Invoke("IsObjectMoving", "i", objectid);
 }
 
-int CreatePlayerObject(int playerid, int modelid, float X, float Y, float Z, float rX, float rY, float rZ, float DrawDistance) {
+int CreatePlayerObject(int playerid, int modelid, double X, double Y, double Z, double rX, double rY, double rZ, double DrawDistance) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("CreatePlayerObject", "iifffffff", playerid, modelid, X, Y, Z, rX, rY, rZ, DrawDistance);
 }
 
-int SetPlayerObjectPos(int playerid, int objectid, float X, float Y, float Z) {
+int SetPlayerObjectPos(int playerid, int objectid, double X, double Y, double Z) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("SetPlayerObjectPos", "iifff", playerid, objectid, X, Y, Z);
 }
 
-int GetPlayerObjectPos(int playerid, int objectid, float* X, float* Y, float* Z) {
+int GetPlayerObjectPos(int playerid, int objectid, double* X, double* Y, double* Z) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("GetPlayerObjectPos", "iiFFF", playerid, objectid, X, Y, Z);
 }
 
-int SetPlayerObjectRot(int playerid, int objectid, float RotX, float RotY, float RotZ) {
+int SetPlayerObjectRot(int playerid, int objectid, double RotX, double RotY, double RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("SetPlayerObjectRot", "iifff", playerid, objectid, RotX, RotY, RotZ);
 }
 
-int GetPlayerObjectRot(int playerid, int objectid, float* RotX, float* RotY, float* RotZ) {
+int GetPlayerObjectRot(int playerid, int objectid, double* RotX, double* RotY, double* RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("GetPlayerObjectRot", "iiFFF", playerid, objectid, RotX, RotY, RotZ);
 }
@@ -124,7 +124,7 @@ int DestroyPlayerObject(int playerid, int objectid) {
   return g_native_function_manager->Invoke("DestroyPlayerObject", "ii", playerid, objectid);
 }
 
-int MovePlayerObject(int playerid, int objectid, float X, float Y, float Z, float Speed, float RotX, float RotY, float RotZ) {
+int MovePlayerObject(int playerid, int objectid, double X, double Y, double Z, double Speed, double RotX, double RotY, double RotZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("MovePlayerObject", "iifffffff", playerid, objectid, X, Y, Z, Speed, RotX, RotY, RotZ);
 }
@@ -139,7 +139,7 @@ int IsPlayerObjectMoving(int playerid, int objectid) {
   return g_native_function_manager->Invoke("IsPlayerObjectMoving", "ii", playerid, objectid);
 }
 
-int AttachPlayerObjectToPlayer(int objectplayer, int objectid, int attachplayer, float OffsetX, float OffsetY, float OffsetZ, float rX, float rY, float rZ) {
+int AttachPlayerObjectToPlayer(int objectplayer, int objectid, int attachplayer, double OffsetX, double OffsetY, double OffsetZ, double rX, double rY, double rZ) {
   CHECK(g_native_function_manager);
   return g_native_function_manager->Invoke("AttachPlayerObjectToPlayer", "iiiffffff", objectplayer, objectid, attachplayer, OffsetX, OffsetY, OffsetZ, rX, rY, rZ);
 }
