+++ src/server/bindings/vehicles.h
@@ -16,7 +16,7 @@
 #ifndef SERVER_BINDINGS_VEHICLES_H_
 #define SERVER_BINDINGS_VEHICLES_H_
 
-// Generated on 2014-09-27.
+// Generated on 2014-09-28.
 // Do not modify by hand, instead, look at /scripts/write_bindings.py.
 namespace samp {
 
@@ -38,15 +38,15 @@ const int VEHICLE_PARAMS_UNSET = -1;
 const int VEHICLE_PARAMS_OFF = 0;
 const int VEHICLE_PARAMS_ON = 1;
 
-int CreateVehicle(int vehicletype, float x, float y, float z, float rotation, int color1, int color2, int respawn_delay);
+int CreateVehicle(int vehicletype, double x, double y, double z, double rotation, int color1, int color2, int respawn_delay);
 int DestroyVehicle(int vehicleid);
 int IsVehicleStreamedIn(int vehicleid, int forplayerid);
-int GetVehiclePos(int vehicleid, float* x, float* y, float* z);
-int SetVehiclePos(int vehicleid, float x, float y, float z);
-int GetVehicleZAngle(int vehicleid, float* z_angle);
-int GetVehicleRotationQuat(int vehicleid, float* w, float* x, float* y, float* z);
-float GetVehicleDistanceFromPoint(int vehicleid, float X, float Y, float Z);
-int SetVehicleZAngle(int vehicleid, float z_angle);
+int GetVehiclePos(int vehicleid, double* x, double* y, double* z);
+int SetVehiclePos(int vehicleid, double x, double y, double z);
+int GetVehicleZAngle(int vehicleid, double* z_angle);
+int GetVehicleRotationQuat(int vehicleid, double* w, double* x, double* y, double* z);
+double GetVehicleDistanceFromPoint(int vehicleid, double X, double Y, double Z);
+int SetVehicleZAngle(int vehicleid, double z_angle);
 int SetVehicleParamsForPlayer(int vehicleid, int playerid, int objective, int doorslocked);
 int ManualVehicleEngineAndLights();
 int SetVehicleParamsEx(int vehicleid, int engine, int lights, int alarm, int doors, int bonnet, int boot, int objective);
@@ -57,8 +57,8 @@ int AddVehicleComponent(int vehicleid, int componentid);
 int RemoveVehicleComponent(int vehicleid, int componentid);
 int ChangeVehicleColor(int vehicleid, int color1, int color2);
 int ChangeVehiclePaintjob(int vehicleid, int paintjobid);
-int SetVehicleHealth(int vehicleid, float health);
-int GetVehicleHealth(int vehicleid, float* health);
+int SetVehicleHealth(int vehicleid, double health);
+int GetVehicleHealth(int vehicleid, double* health);
 int AttachTrailerToVehicle(int trailerid, int vehicleid);
 int DetachTrailerFromVehicle(int vehicleid);
 int IsTrailerAttachedToVehicle(int vehicleid);
@@ -68,9 +68,9 @@ int GetVehicleModel(int vehicleid);
 int GetVehicleComponentInSlot(int vehicleid, int slot);
 int GetVehicleComponentType(int component);
 int RepairVehicle(int vehicleid);
-int GetVehicleVelocity(int vehicleid, float* X, float* Y, float* Z);
-int SetVehicleVelocity(int vehicleid, float X, float Y, float Z);
-int SetVehicleAngularVelocity(int vehicleid, float X, float Y, float Z);
+int GetVehicleVelocity(int vehicleid, double* X, double* Y, double* Z);
+int SetVehicleVelocity(int vehicleid, double X, double Y, double Z);
+int SetVehicleAngularVelocity(int vehicleid, double X, double Y, double Z);
 int GetVehicleDamageStatus(int vehicleid, int* panels, int* doors, int* lights, int* tires);
 int UpdateVehicleDamageStatus(int vehicleid, int panels, int doors, int lights, int tires);
 int SetVehicleVirtualWorld(int vehicleid, int worldid);
